split webpack config into dev and prod
