+++ .gitignore
@@ -2,4 +2,5 @@ node_modules
 .idea
 config/sequelizeCliConfig.js
 config/speechConfig.js
+public/bundle
 server.js
