+++ webpack.client.common.js
@@ -1,0 +1,32 @@
+const Path = require('path');
+const REACT_ROOT = Path.resolve(__dirname, 'react/');
+
+module.exports = {
+  target: 'web',
+  entry : ['babel-polyfill', 'whatwg-fetch', './react/client.js'],
+  output: {
+    path      : Path.join(__dirname, 'public/bundle/'),
+    publicPath: 'public/bundle/',
+    filename  : 'bundle.js',
+  },
+  module: {
+    loaders: [
+      {
+        test   : /.jsx?$/,
+        loader : 'babel-loader',
+        exclude: /node_modules/,
+        query  : {
+          presets: ['es2015', 'react', 'stage-2'],
+        },
+      },
+    ],
+  },
+  resolve: {
+    modules: [
+      REACT_ROOT,
+      'node_modules',
+      __dirname,
+    ],
+    extensions: ['.js', '.jsx', '.scss'],
+  },
+};
